chore: bump revm to v2.2.2, revm-inspectors to v2.2.1, alloy-evm to v2.2.1 (#30)
* chore: bump revm to v2.2.2, revm-inspectors to v2.2.1, alloy-evm to v2.2.1

- Bump mantle-xyz/revm tag v2.2.1 → v2.2.2 (revm, revm-bytecode,
  revm-state, revm-primitives, revm-interpreter, revm-inspector,
  revm-context, revm-context-interface, revm-database,
  revm-database-interface, op-revm)
- Bump mantle-xyz/revm-inspectors tag v2.2.0 → v2.2.1
- Bump mantle-xyz/evm tag v2.2.0 → v2.2.1 (alloy-evm, alloy-op-evm)
- Enable reth-codec in op-primitives default features

* fix: resolve fmt, typos, and deny sources CI failures

- fix(fmt): reorder imports in ethereum/primitives receipt.rs to satisfy
  nightly rustfmt (place __private before Compact per rustfmt ordering)
- fix(typos): correct spelling writeable -> writable in libmdbx-rs env doc
- fix(deny): add mantle-xyz forks to allow-git in deny.toml
  (mantle-xyz/revm, revm-inspectors, evm, op-alloy)

* chore: bump MSRV from 1.88 to 1.91

alloy-*@1.7.3 requires rustc >= 1.91. Update the declared MSRV to match
the actual requirement imposed by the dependency tree.

- Cargo.toml: rust-version 1.88 -> 1.91
- .github/workflows/lint.yml: MSRV toolchain 1.88 -> 1.91
- Dockerfile.reproducible: base image rust:1.88-bookworm -> 1.91-bookworm

Note: Dockerfile.x86_64-pc-windows-gnu intentionally left at 1.88.0
due to xargo incompatibility with Rust >= 1.89 (cross-rs/cross#1701).

Verified: `cargo +1.91 build --bin op-reth` passes.

* fix: resolve clippy, feature-propagation, and deny advisory CI failures

- fix(clippy): use sort_unstable_by_key in reth-trie-common/updates.rs
  (lines 133, 279: sort_unstable_by(|a,b| a.0.cmp(&b.0)) -> sort_unstable_by_key(|a| a.0))
- fix(feature-propagation): propagate js-tracer feature from
  reth-optimism-node to reth-rpc-eth-api (zepter requirement)
- fix(deny): add 8 pre-existing transitive dependency advisories to
  ignore list (RUSTSEC-2025-0141, -0137, RUSTSEC-2026-0007, -0009,
  -0041, -0049, -0067, -0068); all identical to main branch before this PR

* fix(feature-propagation): propagate std/serde features in mantle crates

Auto-fixed by zepter: 14 missing feature propagations introduced when
revm v2.2.2 and alloy-evm v2.2.1 added serde/std features to their
public API. Affected crates:
- crates/mantle-hardforks: std and serde -> alloy-chains, alloy-primitives,
  op-revm, reth-optimism-forks, reth-primitives-traits, revm, etc.
- crates/optimism/chainspec: serde -> reth-mantle-forks
- crates/optimism/consensus: std -> reth-mantle-forks
- crates/optimism/evm: std -> reth-mantle-forks, tracing

* ci: replace Reth runner group with ubuntu-latest

The mantle-xyz/reth fork does not have a 'Reth' self-hosted runner group
configured, causing all jobs using it to fail immediately. Replace with
ubuntu-latest to unblock CI on the fork.

Affected: bench, compact, e2e, hive, integration, kurtosis, kurtosis-op,
prepare-reth, stage, sync-era, sync, unit workflows.

* fix(security): upgrade bytes/lz4_flex/ruint and improve advisory docs

Security fixes (per code review):
- bytes 1.10.1 -> 1.11.1 (RUSTSEC-2026-0007: integer overflow in reserve)
- lz4_flex 0.11.5 -> 0.11.6 (RUSTSEC-2026-0041: memory leak on bad input)
- ruint 1.17.0 -> 1.17.1 (RUSTSEC-2025-0137: UB in reciprocal_mg10)

Remaining ignored advisories (RUSTSEC-2025-0141, -2026-0049, -0067,
-0068, -0009) now have full comments explaining affected crate, risk
context, and why each cannot be upgraded in this PR.

* fix: resolve check-toml, clippy, and docs CI failures

- fix(check-toml): run dprint fmt on 4 Cargo.toml files modified by
  zepter (tabs -> spaces per dprint config)
- fix(clippy): collapse nested if into match guards in chainspec/spec.rs
  (clippy::collapsible_match, lines 732 and 737)
- fix(docs): remove stabilized doc_auto_cfg from reth-mantle-forks feature
  attribute (E0557: feature has been removed; stabilized in Rust 1.82)

* fix: resolve fmt and clippy failures in trie/chainspec/discv4

- fix(fmt): apply nightly rustfmt to chainspec/spec.rs and
  net/discv4/src/lib.rs (match arm guard formatting changed in nightly)
- fix(clippy): replace sort_unstable_by with sort_unstable_by_key in
  trie/trie/src/verify.rs:144 (unnecessary_sort_by lint)
  storage_tries sorted in reverse using std::cmp::Reverse(a.0)

* fix: resolve clippy warnings across multiple crates

Fix all clippy lint failures reported by make lint:
- collapsible_match: collapse if-in-match into match guards (transaction-pool, cli/db/tui)
- unnecessary_sort_by: use sort_by_key (era-downloader, trie, stages)
- manual_is_some_and: .filter().is_some() → .as_ref().is_some_and() (provider, db-common, stages, trie)
- explicit_counter_loop: use range+zip instead of manual counters (pending pool, provider)
- useless_into_iter: remove redundant .into_iter() calls (transaction-pool, stages, trie, examples)
- map_iter_values: use .values()/.into_values() instead of .iter() with ignored key (trie/db, engine/util)
- manual_option_zip: use Option::zip instead of and_then+map (optimism/node)
- useless_cast: remove no-op u64 as u64 casts (ethereum/consensus)

* revert: restore Reth runner group in CI workflows

Reverts the ubuntu-latest change for 12 workflow files. The runner group
issue needs to be solved at infrastructure level (configuring the Reth
runner group in GitHub Settings), not by changing runner type.

lint.yml retains the MSRV 1.88 → 1.91 update as that is a code change.

* fix(deny): remove unnecessary RUSTSEC-2026-0049 ignore

rustls-webpki 0.103.8 (currently in Cargo.lock) already satisfies the
fix requirement (>= 0.102.9). The ignore entry was added under the
assumption that the upstream alloy/revm pin blocked the upgrade, but
the revm v2.2.2 bump in this PR brought rustls-webpki to 0.103.8,
making the whitelist entry obsolete.

* fix(deny): restore RUSTSEC-2026-0049 ignore with correct fix version

Revert the earlier removal. rustls-webpki 0.103.8 is still in the
vulnerable range (>= 0.102.0-alpha.0, < 0.103.10); the 0.103.x fix
requires >= 0.103.10, not >= 0.102.9 as previously noted.

Update comment to reflect the accurate patched version so the next
dep bump knows exactly what to upgrade to.

* fix(deny): correct patched versions in advisory ignore comments

All three advisory comments had inaccurate "Fixed in" versions:
- tar RUSTSEC-2026-0067/0068: fix is >= 0.4.45, not >= 0.4.44
  (0.4.44 is the last *vulnerable* version, not the fix)
- time RUSTSEC-2026-0009: fix is >= 0.3.47, not >= 0.3.37

Also annotate current versions so it's clear these are still pending
upgrade, not already resolved.

+++ crates/ethereum/consensus/src/lib.rs
@@ -213,7 +213,7 @@ mod tests {
     #[test]
     fn test_valid_gas_limit_increase() {
         let parent = header_with_gas_limit(GAS_LIMIT_BOUND_DIVISOR * 10);
-        let child = header_with_gas_limit((parent.gas_limit + 5) as u64);
+        let child = header_with_gas_limit(parent.gas_limit + 5);
 
         assert_eq!(
             validate_against_parent_gas_limit(&child, &parent, &ChainSpec::<Header>::default()),
@@ -228,7 +228,7 @@ mod tests {
 
         assert_eq!(
             validate_against_parent_gas_limit(&child, &parent, &ChainSpec::<Header>::default()),
-            Err(ConsensusError::GasLimitInvalidMinimum { child_gas_limit: child.gas_limit as u64 })
+            Err(ConsensusError::GasLimitInvalidMinimum { child_gas_limit: child.gas_limit })
         );
     }
